restructure github plugin
